requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==23.4
python-decouple==3.8
celery==5.3.4
redis==5.0.1
gunicorn==21.2.0
whitenoise==6.6.0
